Started ripping apart the (bigger) player listener class.
