+++ src/net/jitse/npclib/listeners/PlayerListener.java
@@ -12,27 +12,12 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerMoveEvent;
-import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.player.PlayerTeleportEvent;
 
 /**
  * @author Jitse Boonstra
  */
 public class PlayerListener implements Listener {
-
-    @EventHandler
-    public void onPlayerQuit(PlayerQuitEvent event) {
-        Player player = event.getPlayer();
-        for (NPC npc : NPCManager.getAllNPCs()) {
-            if (npc.getAutoHidden().contains(player.getUniqueId())) {
-                npc.getAutoHidden().remove(player.getUniqueId());
-            }
-
-            if (npc.isActuallyShown(player)) {
-                npc.hide(player);
-            }
-        }
-    }
 
     @EventHandler
     public void onPlayerMove(PlayerMoveEvent event) {
